requirements.txt
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
SQLAlchemy==1.2.12
PyJWT==1.4.0
flask-cors==3.0.7
flask-api==1.0